Text baking and more
Former-commit-id: 21bec3f52b026926e2647ffa337f66f734d6b2aa
Former-commit-id: 141e2d916956f126fd83c3908fb84051a0ab9761

+++ Graphics/Shader/ShaderProgram.h
@@ -99,7 +99,7 @@ namespace TealEngine {
 
 		void setUniform(std::string name, glm::mat4* mat, int size) 
 		{ 
-			float* matv = new float[16 * size];
+			float* matv = new float[long(16 * size)];
 			for (int i = 0; i < size; i++)
 				memcpy(matv + i * 16, value_ptr(mat[i]), sizeof(float) * 16);
 			uniforms[name]->setm4fv(matv, size);
